pyproject.toml
[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "evdemand"
version = "0.1.0"
description = "Spatiotemporal GNN for EV charger demand forecasting and siting"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "torch>=2.2",
    "pandas>=2.1",
    "numpy>=1.26",
    "pyyaml>=6.0",
    "scikit-learn>=1.3",
    "networkx>=3.2",
    "tqdm>=4.66",
]

[tool.setuptools.packages.find]
where = ["src"]
